% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{source.attribution.mcmc.diagnostics}
\alias{source.attribution.mcmc.diagnostics}
\title{MCMC diagnostics for the source attribution algorithm}
\usage{
source.attribution.mcmc.diagnostics(mcmc.file, mc = NULL,
  control = list(burnin.p = 0.2, regex_pars = "*", credibility.interval =
  0.95, pdf.plot.n.worst.case.parameters = 10, pdf.plot.all.parameters =
  FALSE, pdf.height.per.par = 1.2, outfile.base = gsub("\\\\.rda", "",
  mcmc.file)))
}
\arguments{
\item{mcmc.file}{Full file name to MCMC output from function \code{source.attribution.mcmc}}

\item{mc}{MCMC object from function \code{source.attribution.mcmc}; this can be specified alternatively to `mcmc.file`.}

\item{control}{List of input arguments that control the behaviour of the source attribution algorithm:
\itemize{
 \item{"burnin.p"}{Proportion of MCMC iterations that are removed as burn-in period.}
 \item{"regex_pars"}{Regular expression to select the parameter names for which diagnostics are computed. The default is all parameters, which is '*'.}
 \item{"credibility.interval"}{Width of the marginal posterior credibility intervals.}
 \item{"pdf.plot.n.worst.case.parameters"}{Integer which specifies the number of parameters with smallest effective sample size that are inspected in detail. If set to 0, worst case analyses are not performed.}
 \item{"pdf.plot.all.parameters"}{Flag which specifies if traces shall be plotted for all parameters. If set to TRUE, very large pdfs may be created.}
 \item{"pdf.height.per.par"}{Most plots show diagnostics with parameters listed on the y-axis. This value controls the plot height in inches for each free parameter.}
 \item{"outfile.base"}{Start of the full file name for all output files.}
}}
}
\value{
Nothing. Diagnostic plots and csv files are written to disk.
}
\description{
MCMC diagnostics for the source attribution algorithm
}
\author{
Xiaoyue Xi, Oliver Ratmann
}
