% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyquantities.R
\name{source.attribution.mcmc.getKeyQuantities}
\alias{source.attribution.mcmc.getKeyQuantities}
\title{Estimate Flows, Sources, WAIFM, Flow ratios}
\usage{
source.attribution.mcmc.getKeyQuantities(infile = NULL, mc = NULL,
  pars = NULL, dobs = NULL, control = NULL)
}
\arguments{
\item{infile}{Full file name to MCMC output or aggregated MCMC output from function \code{source.attribution.mcmc}}

\item{control}{List of input arguments that control the behaviour of the derivation of key quantities:
\itemize{
 \item{"quantiles"}{Named list of quantiles. Default: c('CL'=0.025,'IL'=0.25,'M'=0.5,'IU'=0.75,'CU'=0.975)}
 \item{"flowratios"}{Cector of length 3. First element: name of flow ratio. Second element: name of transmission pair category for enumerator of flow ratio. Third element: name of transmission pair category for denominator of flow ratio.}
 \item{"outfile"}{Full file name for output csv file.}
}}
}
\value{
If `outfile` is not specified, the output is returned as a data.table. If `outfile` is specified, the output is written to a csv file.
}
\description{
Estimate Flows, Sources, WAIFM, Flow ratios
}
\author{
Xiaoyue Xi, Oliver Ratmann
}
