% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{source.attribution.mcmc}
\alias{source.attribution.mcmc}
\title{Source attribution while adjusting for sampling bias}
\usage{
source.attribution.mcmc(dobs, dprior, control = list(seed = 42, mcmc.n =
  1000, verbose = 1, outfile = "SAMCMCv190327.rda"))
}
\arguments{
\item{dobs}{Data.table of observed number of transmission events for each transmission pair category.}

\item{dprior}{Data.table of draws from the prior distribution of sampling probabilities.}

\item{control}{List of input arguments that control the behaviour of the source attribution algorithm:
\itemize{
 \item{"seed"}{Random number seed, for reproducibility.}
 \item{"mcmc.n"}{Guide on the number of MCMC iterations. The actual number of iterations will be slightly larger, and a multiple of the number of iterations needed to complete one MCMC sweep through all free parameters.}
 \item{"verbose"}{Flag to switch on/off verbose output.}
 \item{"outfile"}{Full path name to which the MCMC output is saved to in RDA format, in an object called \code{mc}.}
}}
}
\value{
If `outfile` is not specified, phyloflows MCMC output is returned, which is a list object. If `outfile` is specified, phyloflows MCMC output is written to an RDA file.
}
\description{
Source attribution while adjusting for sampling bias
}
\author{
Xiaoyue Xi, Oliver Ratmann
}
