% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_stan.R
\name{source.attribution.mcmc.stan}
\alias{source.attribution.mcmc.stan}
\title{Source attribution while adjusting for sampling bias}
\usage{
source.attribution.mcmc.stan(
  dobs,
  dprior.fit,
  control = list(seed = 42, mcmc.n = 1000, method = "gamma", outfile = "StanMCMC.rda")
)
}
\arguments{
\item{dobs}{Data.table of observed number of transmission events for each transmission pair category.}

\item{control}{List of input arguments that control the behaviour of the source attribution algorithm:
\itemize{
 \item{"seed"}{Random number seed, for reproducibility.}
 \item{"mcmc.n"}{Guide on the number of MCMC iterations. The actual number of iterations will be slightly larger, and a multiple of the number of iterations needed to complete one MCMC sweep through all free parameters.}
 \item{"method"}{Choose flow prior among 'gamma'(default) and 'gp'.}
 \item{"outfile"}{Full path name to which the MCMC output is saved to in RDA format, in an object called \code{fit}.}
}}

\item{dprior}{Data.table of parameters from the prior distribution of sampling probabilities.}
}
\value{
If `outfile` is not specified, phyloflows MCMC output is returned, which is a list object. If `outfile` is specified, phyloflows MCMC output is written to an RDA file.
}
\description{
Source attribution while adjusting for sampling bias
}
\examples{
require(data.table)
require(phyloflows)
# load input data set
data(twoGroupFlows1, package="phyloflows")
dobs <- twoGroupFlows1$dobs
dprior <- twoGroupFlows1$dprior

#
# run MCMC and return output to R
# this is usually fine for simple source attribution models
# that don t take long to run
#
\dontrun{
control <- list(seed=42,    # seed for reproducibility
                mcmc.n=5e4, # guide on the number of MCMC iterations, the actual number of MCMC iterations may be a little bit larger
                method='gamma'   # switch on for output to each MCMC iteration
                )
mc <- phyloflows:::source.attribution.mcmc.stan(dobs, dprior, control)
}

#
# run MCMC and save output to file
# this is recommended when the source attribution is more complex
# so you can deploy the MCMC on a high performance cluster
#
\dontrun{
mcmc.file <- 'mcmcm_output.rda'
control <- list(seed=42,    # seed for reproducibility
                mcmc.n=5e4, # guide on the number of MCMC iterations, the actual number of MCMC iterations may be a little bit larger
                 method='gamma',   # switch on for output to each MCMC iteration
                outfile= mcmc.file # store MCMC output as rda file here
                )
mc <- phyloflows:::source.attribution.mcmc.stan(dobs, dprior, control)
}

}
\seealso{
\link{\code{source.attribution.mcmc.diagnostics}}, \link{\code{source.attribution.mcmc.aggregateToTarget}}, \link{\code{source.attribution.mcmc.getKeyQuantities}}
}
\author{
Xiaoyue Xi, Oliver Ratmann
}
