% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_and_diagnostics.R
\name{source.attribution.mcmc.aggregateToTarget}
\alias{source.attribution.mcmc.aggregateToTarget}
\title{Aggregate MCMC output to target parameters}
\usage{
source.attribution.mcmc.aggregateToTarget(mcmc.file, daggregateTo,
  control = list(burnin.p = NA_real_, thin = NA_integer_, regex_pars =
  "*", outfile = gsub("\\\\.rda", "_aggregated.csv", mcmc.file)))
}
\arguments{
\item{mcmc.file}{Full file name to MCMC output from function \code{source.attribution.mcmc}}

\item{daggregateTo}{Data.table that maps the categories of transmission pairs used in the MCMC to lower-dimensional categories that are of primary interest.}

\item{control}{List of input arguments that control the behaviour of the source attribution algorithm:
\itemize{
 \item{"burnin.p"}{Proportion of MCMC iterations that are removed as burn-in period.}
 \item{"thin"}{Thin MCMC output to every nth MCMC iteration.}
 \item{"regex_pars"}{Regular expression to select the parameters that are to be aggregated.}
 \item{"outfile"}{Full file name of the output csv file.}
}}
}
\value{
NULL. Monte Carlo samples are written to a csv file.
}
\description{
Aggregate MCMC output to target parameters
}
